requirements.txt
absl-py==2.3.1
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
immutabledict==4.2.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
narwhals==2.0.1
numpy==2.3.2
openpyxl==3.1.5
ortools==9.14.6206
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
fastapi==0.115.6
sqlmodel==0.0.19
uvicorn==0.30.6
alembic==1.14.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.9.0
pydantic-settings==2.7.0
email-validator==2.2.0
httpx==0.27.2
